Add shorthand to jump to shared VirtualBox directory
In the emulated versions of Windows I run (for testing websites on IE or
scripts with COMMAND.COM/CMD.exe), the host-to-guest folder is mapped to
a virtual network drive of "G:\".

Since I'm already used to navigating to that directory in Windows with a
two-character command, it makes sense to have the keep the same shortcut
in Bash.