Fix `checkbashisms` override and add `map` alias
Credit for the latter goes to mathiasbynens/dotfiles@5400992. The former
addresses something I forgot to remove when overriding the checkbashisms
command. It was aliased to `perl /usr/local/bin/checkbashisms --`, but I
realising clearing the PERL5LIB variable was a better way to do it. Only
I forgot to delete the double-dash at the end of the alias.

Like many commit messages authored when sleep-deprived, the verbosity of
this one is primarily fuelled by an obsession for neatly lining up lines
of text at the 72-character margin.