Include "missing" manual-page directories on macOS
We give these directories lowest priority in $MANPATH, as they're likely
to contain the least up-to-date information (macOS ships with antiques).
However, this does bring back pages for make(1), strings(1), and others
that weren't located in any of the "standard" directories.

These paths are all Darwin-specific, so we don't bother statting them on
other platforms.