Make prompt-string available in other shells
Tested with ksh(1), dash(1), mksh(1) and zsh(1). OpenBSD's wscons(4) has
been exempted from fancy-looking prompts, since it clashes with the raw,
old-school feel of a pure-text interface. That, and I'm already too used
to seeing it without colour, heh.

Tried testing `prompt.sh` with posh(1), but the version for macOS failed
to start as a login shell, and was unable to run even simple commands in
an interactive session. Something's seriously wrong with it, or at least
with Homebrew's build of it.