Add a transcription of my ancient PC's BIOS screen
Trying to boot from a USB drive usually involves entering the boot menu,
selecting the device, and hoping my pre-Cambrian motherboard understands
whatever's responsible for bootstrapping an OS. That first step requires
me to smash one of 12 function-keys on my keyboard, the name of which is
displayed at startup for only a split-second (and sometimes not at all).

This file loosely resembles my BIOS splash screen in plain-text form. It
was originally saved to PasteBin (https://pastebin.com/KYMhnzFw), but my
dotfiles are probably a better place to store it.