Add shortcut to livecheck `alhadis/homebrew-troff`
At the moment, `brew lc` is inadequate for livechecking a particular tap
due to `~/.brew_livecheck_watchlist` only supporting casks and formulae:
entries such as "--tap alhadis/troff" are only met with an error.

Of course, I could maintain a list of formula names, but that's a PITFA,
and I'd rather not bust out Watchman to monitor the tap for changes.