Hide Wikipedia's contest promotion banner
Dismissing the banner doesn't work when navigating between subdomains of
`*.wikipedia.org`, and this presumably has something to do with a cookie
not being set properly.

However, this is giving me the shits as I hop between Wikipedia projects
to update stale thumbnails, so I really have no choice in the matter.