Fix osascript(1)'s warning over "schema-less" URLs
The `info for` command coerces its argument to a File value or an Alias.
If the subject isn't an absolute POSIX path, this warning is displayed:

    osascript[15801:5000540] CFURLGetFSRef was passed a URL which
    has no scheme (the URL will not work with other CFURL routines)

Both Files and Aliases use Classic MacOS paths for specifying locations,
hence the added confusion over the interpretation of Unix-style paths.

Source: https://stackoverflow.com/a/12573326