Add the NVD to Firefox's search engines
Since the database is hosted on a US government website, assume it'll be
driven into the ground by the incompetence of the current administration
sooner or later. If that happens, replace this entry with CVE.org:

    Name: Common Vulnerabilities and Exposures
    URL:  https://www.cve.org/CVERecord/SearchResults?query=%s
    Icon: https://www.cve.org/cvePurpleVFavicon.svg

This semi-serious contingency assumes CVE.org remains accessible despite
whatever atavism that took out America's `.gov`-hosted domains.